Move all into the plugins folder
