+++ public/main/plugins/coincap/index.js
