+++ public/main/plugins/ethWallet/block.js
