+++ public/main/plugins/ethWallet/db.js
