+++ public/main/plugins/ethWallet/eth.js
