+++ public/main/plugins/ethWallet/key.js
