+++ public/main/plugins/ethWallet/send.js
@@ -1,6 +1,6 @@
 const web3 = require('web3')
-const { decrypt: decryptOld } = require('../crypto/aes256cbc')
-const { encrypt, decrypt } = require('../crypto/aes256cbcIv')
+const { decrypt: decryptOld } = require('../../crypto/aes256cbc')
+const { encrypt, decrypt } = require('../../crypto/aes256cbcIv')
 
 const {
   completedTransactionParams,
