+++ public/main/plugins/ethWallet/settings.js
