+++ public/main/plugins/ethWallet/tracesParser.js
