+++ public/main/plugins/ethWallet/transactionParser.js
