+++ public/main/plugins/ethWallet/wallet.js
