+++ public/main/plugins/ethWallet/web3.js
