+++ public/main/plugins/metronome/auctions.js
