+++ public/main/plugins/metronome/converter.js
