+++ public/main/plugins/metronome/index.js
