+++ public/main/plugins/metronome/settings.js
