+++ public/main/plugins/metronome/transactionParser.js
