+++ public/main/plugins/tokens/settings.js
