+++ public/main/plugins/tokens/transactionParser.js
